refactor(types): move backend models to shared package
