+++ apps/backend/src/routes/about.ts
@@ -1,9 +1,10 @@
 import { Elysia } from "elysia"
-import { AboutResponse } from "../models/about.model"
+import { AboutResponse } from "@area/types"
 import { serviceRegistry } from "../services/registry"
 
-export const aboutRoute = new Elysia({ prefix: "" })
-  .get("/about.json", ({ request }) => {
+export const aboutRoute = new Elysia({ prefix: "" }).get(
+  "/about.json",
+  ({ request }) => {
     const clientIp = request.headers.get("x-forwarded-for") || request.headers.get("x-real-ip") || "unknown"
     return {
       client: { host: clientIp },
