+++ apps/backend/src/routes/areas.ts
@@ -2,7 +2,16 @@ import { Elysia, t } from "elysia"
 import { authMiddleware } from "../middlewares/better-auth"
 import { prisma } from "../db/prisma"
 import { serviceRegistry } from "../services/registry"
-import { AreaErrorResponse, AreaResponse, AreasListResponse, AreaStatsResponse, AreaToggleResponse, CreateAreaBody, ListAreasQuery, UpdateAreaBody } from "../models/area.model"
+import { 
+  AreaErrorResponse, 
+  AreaResponse, 
+  AreasListResponse, 
+  AreaStatsResponse, 
+  AreaToggleResponse, 
+  CreateAreaBody, 
+  ListAreasQuery, 
+  UpdateAreaBody 
+} from "@area/types"
 
 export const areasRoutes = new Elysia({ prefix: "/api/areas" })
   .use(authMiddleware)
